Fix for mouse popup bug (wrong feature selected for popup text)
git-svn-id: https://igv.googlecode.com/svn/trunk@421 17392f64-ead8-4cea-ae29-09b3ab513800

+++ src/org/broad/igv/track/FeatureCollectionSource.java
@@ -329,7 +329,7 @@ public class FeatureCollectionSource implements FeatureSource {
 
             // give a 2 pixel window, otherwise very narrow features will be missed.
             double bpPerPixel = frame.getScale();
-            double minWidth = 2 * bpPerPixel;    /* * */
+            int minWidth = (int) (2 * bpPerPixel);    /* * */
 
             if (scores == null) {
                 return "";
